requirements.txt
streamlit==1.38.0
Pillow==10.4.0
requests==2.32.3
streamlit-image-comparison==0.0.4
